+++ lib/ui/widgets/auth/auth_widget.dart
@@ -1,296 +1,163 @@
-// import 'dart:math';
-//
-// import 'package:comics_db_app/app_colors.dart';
-// import 'package:comics_db_app/library/widgets/inherited/notifier_provider.dart';
-// import 'package:comics_db_app/resources/resources.dart';
-// import 'package:comics_db_app/ui/widgets/auth/auth_view_cubit_state.dart';
-// import 'package:flutter/material.dart';
-//
-// class AuthWidget extends StatefulWidget {
-//   const AuthWidget({Key? key}) : super(key: key);
-//
-//   @override
-//   State<AuthWidget> createState() => _AuthWidgetState();
-// }
-//
-// class _AuthWidgetState extends State<AuthWidget> with SingleTickerProviderStateMixin {
-//   // If we want to show SignUp
-//   bool _isShowSignUp = false;
-//   late AnimationController _animationController;
-//   late Animation<double> _animationTextRotate;
-//
-//   void setUpAnimation() {
-//     _animationController = AnimationController(vsync: this, duration: AppColors.defaultDuration);
-//     _animationTextRotate = Tween<double>(begin: 0, end: 90).animate(_animationController);
-//   }
-//
-//   void updateView() {
-//     setState(() {
-//       _isShowSignUp = !_isShowSignUp;
-//       _isShowSignUp ? _animationController.forward() : _animationController.reverse();
-//     });
-//   }
-//
-//   @override
-//   void initState() {
-//     setUpAnimation();
-//     super.initState();
-//   }
-//
-//   @override
-//   void dispose() {
-//     _animationController.dispose();
-//     super.dispose();
-//   }
-//
-//   @override
-//   Widget build(BuildContext context) {
-//     // final model = AuthProvider.read(context)?.model;
-//     final model = NotifierProvider.read<AuthModel>(context);
-//     final _size = MediaQuery.of(context).size;
-//     return Scaffold(
-//       resizeToAvoidBottomInset: false,
-//       body: AnimatedBuilder(
-//         animation: _animationController,
-//         builder: (context, _) {
-//           return Stack(
-//             // padding: const EdgeInsets.symmetric(horizontal: 16.0),
-//             children: [
-//               // Login Form
-//               AnimatedPositioned(
-//                 duration: AppColors.defaultDuration,
-//                 // use 88% for login
-//                 width: _size.width * 0.88,
-//                 height: _size.height,
-//                 left: _isShowSignUp ?  - _size.width * 0.76 : 0,
-//                 child: Container(
-//                   color: AppColors.loginBackground,
-//                   child: loginFormColumnPadding(context, model),
-//                 ),
-//               ),
-//               // SignUp
-//               AnimatedPositioned(
-//                 duration: AppColors.defaultDuration,
-//                 height: _size.height,
-//                   width: _size.width * 0.88,
-//                   left:  _isShowSignUp ?_size.width * 0.12 : _size.width * 0.88,
-//                   child: Container(
-//                     color: AppColors.signUpBackground,
-//                     // change to SignUpForm
-//                     child: loginFormColumnPadding(context, model),
-//                   ),
-//               ),
-//               // LOGO
-//               AnimatedPositioned(
-//                 duration: AppColors.defaultDuration,
-//                 top: _size.height * 0.1,
-//                   left: 0,
-//                   right: _isShowSignUp ? - _size.width * 0.06 : _size.width * 0.06,
-//                   child: SizedBox(
-//                     height: 100,
-//                     width: 100,
-//                     child: AnimatedSwitcher(
-//                       duration: AppColors.defaultDuration,
-//                       child: _isShowSignUp ? Image.asset(AppImages.logo) : Image.asset(AppImages.logo),
-//                     ),
-//                   ),
-//               ),
-//               // SignIn animation
-//               AnimatedPositioned(
-//                 duration: AppColors.defaultDuration,
-//                 bottom: _isShowSignUp ? _size.height / 2 -55 : _size.height * 0.15,
-//                   left: _isShowSignUp ? 0 : _size.width * 0.44 - 55,
-//                   child: AnimatedDefaultTextStyle(
-//                     duration: AppColors.defaultDuration,
-//                     textAlign: TextAlign.center,
-//                     style: TextStyle(
-//                         fontSize: _isShowSignUp ? 20 : 30,
-//                         fontWeight: FontWeight.bold,
-//                         color: _isShowSignUp ? Colors.white : Colors.white70),
-//                     child: Transform.rotate(
-//                       angle: -_animationTextRotate.value * pi / 180,
-//                       alignment: Alignment.topLeft,
-//                       child: InkWell(
-//                         onTap: () {
-//                           if (_isShowSignUp) {
-//                             updateView();
-//                           } else {
-//                             // LOGIN
-//                           }
-//                         },
-//                         child: Container(
-//                           padding: const EdgeInsets.symmetric(vertical: AppColors.defaultPadding * 0.75),
-//                           width: 110,
-//                           // color: Colors.red,
-//                           child: Text(
-//                             'Войти'.toUpperCase(),
-//                           ),
-//                         ),
-//                       ),
-//                     ),
-//                   ),
-//               ),
-//               AnimatedPositioned(
-//                 duration: AppColors.defaultDuration,
-//                 bottom: !_isShowSignUp ? _size.height / 2 - 55 : _size.height * 0.15,
-//                 right: _isShowSignUp ? _size.width * 0.44 - 55 : 0,
-//                 child: AnimatedDefaultTextStyle(
-//                   duration: AppColors.defaultDuration,
-//                   textAlign: TextAlign.center,
-//                   style: TextStyle(
-//                       fontSize: !_isShowSignUp ? 20 : 24,
-//                       fontWeight: FontWeight.bold,
-//                       color: _isShowSignUp ? Colors.white : Colors.white70),
-//                   child: Transform.rotate(
-//                     angle: (90 - _animationTextRotate.value) * pi / 180,
-//                     alignment: Alignment.topRight,
-//                     child: InkWell(
-//                       onTap: () {
-//                         if (_isShowSignUp) {
-//                           // SignUp
-//                         } else {
-//                           updateView();
-//                         }
-//                       },
-//                       child: Container(
-//                         padding: const EdgeInsets.symmetric(vertical: AppColors.defaultPadding * 0.75),
-//                         width: 170,
-//                         // color: Colors.red,
-//                         child: Text(
-//                           'Регистрация'.toUpperCase(),
-//                         ),
-//                       ),
-//                     ),
-//                   ),
-//                 ),
-//               ),
-//             ],
-//             // child: loginFormColumnPadding(context, model),
-//           );
-//         }
-//       ),
-//     );
-//   }
-//
-//   Padding loginFormColumnPadding(BuildContext context, AuthModel? model) {
-//     return Padding(
-//         padding: EdgeInsets.symmetric(horizontal: MediaQuery.of(context).size.width * 0.13),
-//         child: Column(
-//           mainAxisAlignment: MainAxisAlignment.center,
-//           children: [
-//             Column(
-//               children: [
-//                 const ErrorMessageWidget(),
-//                 TextFormField(
-//                   controller: model?.loginTextController,
-//                   autofocus: true,
-//                   decoration: formFieldInputDecoration('Почта'),
-//                 ),
-//                 const SizedBox(height: 10),
-//                 // buildPasswordFormField(),
-//                 TextFormField(
-//                   controller: model?.passwordTextController,
-//                   obscureText: true,
-//                   decoration: formFieldInputDecoration('Пароль'),
-//                 ),
-//               ],
-//             ),
-//             const SizedBox(height: 5.0),
-//             resetPasswordRow(),
-//             const SizedBox(height: 20.0),
-//             // loginButton(),
-//             const AuthButtonWidget(),
-//           ],
-//         ),
-//       );
-//   }
-//
-//   InputDecoration formFieldInputDecoration(String labelText) {
-//     return InputDecoration(
-//         contentPadding: const EdgeInsets.symmetric(horizontal: 10.0),
-//         labelStyle: const TextStyle(
-//           color: Colors.white,
-//         ),
-//         // border: const OutlineInputBorder(),
-//         border: InputBorder.none,
-//         hintStyle: const TextStyle(color: Colors.white),
-//         // focusedBorder: const OutlineInputBorder(
-//         //   borderSide: BorderSide(color: AppColors.kPrimaryColor),
-//         // ),
-//         filled: true,
-//         labelText: labelText,
-//         fillColor: Colors.white38);
-//   }
-//
-//   Row resetPasswordRow() {
-//     return Row(
-//       mainAxisAlignment: MainAxisAlignment.end,
-//       children: [
-//         InkWell(
-//           onTap: () {},
-//           child: const Text(
-//             'Забыли пароль?',
-//             style: TextStyle(color: Colors.white),
-//           ),
-//         ),
-//       ],
-//     );
-//   }
-// }
-//
-// class AuthButtonWidget extends StatelessWidget {
-//   const AuthButtonWidget({Key? key}) : super(key: key);
-//
-//   @override
-//   Widget build(BuildContext context) {
-//     // final model = Provider.of<AuthModel>(context, listen: true);
-//     final model = NotifierProvider.watch<AuthModel>(context);
-//     final onPressed =
-//         model?.canStartAuth == true ? () => model?.auth(context) : null;
-//     final child = model?.isAuthProgress == true
-//         ? const CircularProgressIndicator(strokeWidth: 5.0)
-//         : const Text(
-//             'Войти',
-//             style: TextStyle(fontSize: 24, color: Colors.black),
-//           );
-//     return SizedBox(
-//       width: double.infinity,
-//       height: 65,
-//       child: ElevatedButton(
-//         onPressed: onPressed,
-//         child: child,
-//         style: ButtonStyle(
-//           shape: MaterialStateProperty.all(RoundedRectangleBorder(
-//               borderRadius: BorderRadius.circular(30.0))),
-//           side: MaterialStateProperty.all(
-//               const BorderSide(color: AppColors.kPrimaryColorOld)),
-//           backgroundColor: MaterialStateProperty.all(Colors.white),
-//           foregroundColor: MaterialStateProperty.all(AppColors.kPrimaryColorOld),
-//           padding: MaterialStateProperty.all(
-//             const EdgeInsets.symmetric(horizontal: 130.0, vertical: 15.0),
-//           ),
-//         ),
-//       ),
-//     );
-//   }
-// }
-//
-// class ErrorMessageWidget extends StatelessWidget {
-//   const ErrorMessageWidget({Key? key}) : super(key: key);
-//
-//   @override
-//   Widget build(BuildContext context) {
-//     final errorMessage = NotifierProvider.watch<AuthModel>(context)?.errorMessage;
-//     if (errorMessage == null) return const SizedBox.shrink();
-//     return Padding(
-//       padding: const EdgeInsets.only(bottom: 20.0),
-//       child: Text(
-//         errorMessage,
-//         style: const TextStyle(
-//           fontSize: 17,
-//           color: Colors.red,
-//         ),
-//       ),
-//     );
-//   }
-// }
+import 'package:comics_db_app/domain/blocs/auth/auth_view_cubit_state.dart';
+import 'package:comics_db_app/domain/services/auth_data_storage.dart';
+import 'package:comics_db_app/ui/navigation/main_navigation.dart';
+import 'package:comics_db_app/ui/widgets/auth/auth_view_cubit.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+class AuthWidget extends StatelessWidget {
+  const AuthWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocListener<AuthViewCubit, AuthViewCubitState>(
+      listener: _onAuthViewCubitStateChange,
+      child: Provider(
+        create: (_) => AuthDataStorage(),
+        child: Scaffold(
+          appBar: AppBar(
+            title: const Text('Login to your account'),
+          ),
+          body: ListView(
+            children: const [
+              HeaderWidget(),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  void _onAuthViewCubitStateChange(BuildContext context, AuthViewCubitState state) {
+    if (state is AuthViewCubitSuccessAuthState) {
+      MainNavigation.resetNavigation(context);
+    }
+  }
+}
+
+
+
+class FormWidget extends StatelessWidget {
+  FormWidget({Key? key}) : super(key: key);
+
+  final loginTextController = TextEditingController();
+  final passwordTextController = TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    final authDataStorage = context.read<AuthDataStorage>();
+    const textStyle = TextStyle(
+      fontSize: 16,
+      color: Color(0xFF212529),
+    );
+    const textFieldDecorator = InputDecoration(
+      border: OutlineInputBorder(),
+      contentPadding: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
+      isCollapsed: true,
+      fillColor: Colors.red,
+      focusColor: Colors.red,
+      hoverColor: Colors.red,
+    );
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        const _ErrorMessageWidget(),
+        const Text(
+          'Username',
+          style: textStyle,
+        ),
+        const SizedBox(height: 5),
+        TextField(
+          decoration: textFieldDecorator,
+          onChanged: (text) => authDataStorage.login = text,
+        ),
+        const SizedBox(height: 20),
+        const Text(
+          'Password',
+          style: textStyle,
+        ),
+        const SizedBox(height: 5),
+        TextField(
+          decoration: textFieldDecorator,
+          onChanged: (text) => authDataStorage.password = text,
+          obscureText: true,
+        ),
+        const SizedBox(height: 25),
+        Row(
+          children: [
+            const _AuthButtonWidget(),
+            const SizedBox(width: 30),
+            TextButton(
+              onPressed: () {},
+              // style: AppButtonStyle.linkButton,
+              child: const Text('Reset password'),
+            ),
+          ],
+        )
+      ],
+    );
+  }
+}
+
+class _AuthButtonWidget extends StatelessWidget {
+  const _AuthButtonWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final cubit = context.watch<AuthViewCubit>();
+    final authDataStorage = context.read<AuthDataStorage>();
+    final canStartAuth = cubit.state is AuthViewCubitFormFillInProgressState || cubit.state is AuthViewCubitErrorState;
+    const color = Color(0xFF01B4E4);
+    final onPressed =
+        canStartAuth ? () => cubit.auth(login: authDataStorage.login, password: authDataStorage.password) : null;
+    final child = cubit.state is AuthViewCubitAuthProgressState
+        ? const SizedBox(
+            width: 15,
+            height: 15,
+            child: CircularProgressIndicator(strokeWidth: 2),
+          )
+        : const Text('Login');
+    return ElevatedButton(
+      onPressed: onPressed,
+      style: ButtonStyle(
+        backgroundColor: MaterialStateProperty.all(color),
+        foregroundColor: MaterialStateProperty.all(Colors.white),
+        textStyle: MaterialStateProperty.all(
+          const TextStyle(fontSize: 16, fontWeight: FontWeight.w700),
+        ),
+        padding: MaterialStateProperty.all(
+          const EdgeInsets.symmetric(
+            horizontal: 15,
+            vertical: 8,
+          ),
+        ),
+      ),
+      child: child,
+    );
+  }
+}
+
+class _ErrorMessageWidget extends StatelessWidget {
+  const _ErrorMessageWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final errorMessage = context.select((AuthViewCubit c) {
+      final state = c.state;
+      return state is AuthViewCubitErrorState ? state.errorMessage : null;
+    });
+    if (errorMessage == null) return const SizedBox.shrink();
+
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 20),
+      child: Text(
+        errorMessage,
+        style: const TextStyle(
+          fontSize: 17,
+          color: Colors.red,
+        ),
+      ),
+    );
+  }
+}
