--- lib/ui/widgets/auth/auth_widget_simple.dart
@@ -1,205 +1,0 @@
-import 'package:comics_db_app/domain/blocs/auth/auth_view_cubit_state.dart';
-import 'package:comics_db_app/domain/services/auth_data_storage.dart';
-import 'package:comics_db_app/ui/navigation/main_navigation.dart';
-import 'package:comics_db_app/ui/widgets/auth/auth_view_cubit.dart';
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-
-class AuthWidget extends StatelessWidget {
-  const AuthWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocListener<AuthViewCubit, AuthViewCubitState>(
-      listener: _onAuthViewCubitStateChange,
-      child: Provider(
-        create: (_) => AuthDataStorage(),
-        child: Scaffold(
-          appBar: AppBar(
-            title: const Text('Login to your account'),
-          ),
-          body: ListView(
-            children: const [
-              _HeaderWidget(),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  void _onAuthViewCubitStateChange(BuildContext context, AuthViewCubitState state) {
-    if (state is AuthViewCubitSuccessAuthState) {
-      MainNavigation.resetNavigation(context);
-    }
-  }
-}
-
-class _HeaderWidget extends StatelessWidget {
-  const _HeaderWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    const textStyle = TextStyle(
-      fontSize: 16,
-      color: Colors.black,
-    );
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          const SizedBox(height: 25),
-          _FormWidget(),
-          const SizedBox(height: 25),
-          const Text(
-            'In order to use the editing and rating capabilities of TMDb, as well as get personal recommendations you will need to login to your account. If you do not have an account, registering for an account is free and simple.',
-            style: textStyle,
-          ),
-          const SizedBox(height: 5),
-          TextButton(
-            // style: AppButtonStyle.linkButton,
-            onPressed: () {},
-            child: const Text('Register'),
-          ),
-          const SizedBox(height: 25),
-          const Text(
-            'If you signed up but did not get your verification email.',
-            style: textStyle,
-          ),
-          const SizedBox(height: 5),
-          TextButton(
-            // style: AppButtonStyle.linkButton,
-            onPressed: () {},
-            child: const Text('Verify email'),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _FormWidget extends StatelessWidget {
-  _FormWidget({Key? key}) : super(key: key);
-
-  final loginTextController = TextEditingController();
-  final passwordTextController = TextEditingController();
-
-  @override
-  Widget build(BuildContext context) {
-    final authDataStorage = context.read<AuthDataStorage>();
-    const textStyle = TextStyle(
-      fontSize: 16,
-      color: Color(0xFF212529),
-    );
-    const textFieldDecorator = InputDecoration(
-      border: OutlineInputBorder(),
-      contentPadding: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
-      isCollapsed: true,
-      fillColor: Colors.red,
-      focusColor: Colors.red,
-      hoverColor: Colors.red,
-    );
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        const _ErrorMessageWidget(),
-        const Text(
-          'Username',
-          style: textStyle,
-        ),
-        const SizedBox(height: 5),
-        TextField(
-          decoration: textFieldDecorator,
-          onChanged: (text) => authDataStorage.login = text,
-        ),
-        const SizedBox(height: 20),
-        const Text(
-          'Password',
-          style: textStyle,
-        ),
-        const SizedBox(height: 5),
-        TextField(
-          decoration: textFieldDecorator,
-          onChanged: (text) => authDataStorage.password = text,
-          obscureText: true,
-        ),
-        const SizedBox(height: 25),
-        Row(
-          children: [
-            const _AuthButtonWidget(),
-            const SizedBox(width: 30),
-            TextButton(
-              onPressed: () {},
-              // style: AppButtonStyle.linkButton,
-              child: const Text('Reset password'),
-            ),
-          ],
-        )
-      ],
-    );
-  }
-}
-
-class _AuthButtonWidget extends StatelessWidget {
-  const _AuthButtonWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final cubit = context.watch<AuthViewCubit>();
-    final authDataStorage = context.read<AuthDataStorage>();
-    final canStartAuth = cubit.state is AuthViewCubitFormFillInProgressState || cubit.state is AuthViewCubitErrorState;
-    const color = Color(0xFF01B4E4);
-    final onPressed =
-        canStartAuth ? () => cubit.auth(login: authDataStorage.login, password: authDataStorage.password) : null;
-    final child = cubit.state is AuthViewCubitAuthProgressState
-        ? const SizedBox(
-            width: 15,
-            height: 15,
-            child: CircularProgressIndicator(strokeWidth: 2),
-          )
-        : const Text('Login');
-    return ElevatedButton(
-      onPressed: onPressed,
-      style: ButtonStyle(
-        backgroundColor: MaterialStateProperty.all(color),
-        foregroundColor: MaterialStateProperty.all(Colors.white),
-        textStyle: MaterialStateProperty.all(
-          const TextStyle(fontSize: 16, fontWeight: FontWeight.w700),
-        ),
-        padding: MaterialStateProperty.all(
-          const EdgeInsets.symmetric(
-            horizontal: 15,
-            vertical: 8,
-          ),
-        ),
-      ),
-      child: child,
-    );
-  }
-}
-
-class _ErrorMessageWidget extends StatelessWidget {
-  const _ErrorMessageWidget({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final errorMessage = context.select((AuthViewCubit c) {
-      final state = c.state;
-      return state is AuthViewCubitErrorState ? state.errorMessage : null;
-    });
-    if (errorMessage == null) return const SizedBox.shrink();
-
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 20),
-      child: Text(
-        errorMessage,
-        style: const TextStyle(
-          fontSize: 17,
-          color: Colors.red,
-        ),
-      ),
-    );
-  }
-}
