+++ lib/ui/widgets/auth/components/header_widget.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+
+class HeaderWidget extends StatelessWidget {
+  const HeaderWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    const textStyle = TextStyle(
+      fontSize: 16,
+      color: Colors.black,
+    );
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const SizedBox(height: 25),
+          _FormWidget(),
+          const SizedBox(height: 25),
+          const Text(
+            'In order to use the editing and rating capabilities of TMDb, as well as get personal recommendations you will need to login to your account. If you do not have an account, registering for an account is free and simple.',
+            style: textStyle,
+          ),
+          const SizedBox(height: 5),
+          TextButton(
+            // style: AppButtonStyle.linkButton,
+            onPressed: () {},
+            child: const Text('Register'),
+          ),
+          const SizedBox(height: 25),
+          const Text(
+            'If you signed up but did not get your verification email.',
+            style: textStyle,
+          ),
+          const SizedBox(height: 5),
+          TextButton(
+            // style: AppButtonStyle.linkButton,
+            onPressed: () {},
+            child: const Text('Verify email'),
+          ),
+        ],
+      ),
+    );
+  }
+}
